move map into separate component
