+++ client/app/css/singletrip.css
@@ -788,7 +788,7 @@
     flex-direction: column;
     align-items: center; 
     justify-content: center;
-    box-shadow: 0px 4px 8px rgba(0, 0, 0, 0.1);
+    box-shadow: 0px 4px 8px rgba(0, 0, 0, 0.2);
 }
 
 .circle-card:nth-child(2) {
